(maint) There can be only one test DB config.
Remove the modification of the test-db config performed in the fixture
that resets the test db, since it's confusing & error-prone to modify
the DB config with environment variable values in two different places,
and isn't necessary.

Change the name of the environment variables used to set the DB
configuration to `JDBCUTIL_DB*`, where `*` is one of `NAME`, `USER`, or
`PASS`.

Update the travis CI config to use the new environment variables.