Add a retrying pool wrapper
This adds a wrapper for a connection pool that waits for the database to
become available and then runs an initialization function. Any attempts to get
a connection before the init funciton has run will wait for a configurable
timeout and then return a SQLTransientConnectionException (as getConnection in
Hikari would for a normal timeout).