(maint) Use instance? over isa?, fn over #().
Change `convert-result-arrays` to use instance? instead of isa? when
checking if the result value is a java.sql.Array, since isa?  also
checks the clojure inheritance hierarchy, which is not necessary in this
case. It's also a little cleaner since it doesn't require calling
`class` on the value as well. Also change the the anonymous function
literal to just an anonymous function, since we prefer the latter.