(PE-15626) Fix reconcile-sequence-for-column!
Prior to this commit, `reconcile-sequence-for-column!` would cause an
error when given an empty column.

Specifically, it would attempt to set the sequence to 0, which is an
illegal value. As of this commit, it will instead set the sequence to 1,
passing an additional parameter that signals that the next value emitted
should be the current value (i.e., post-increment instead of
pre-increment).