(maint) Refine/de-nilify migration spec
Previously, the output of spec->migration-db-spec could:

1. contain nil values for the :user and/or :password keys
2. match the :password value with the :migration-user value (or various
other permutations)

Neither of these is desirable. In practical usage, passwords are not
always supplied, in which case the key should not be present in the
final map. Furthermore, it should be fine to use passwords for
migrations but not for normal use, or vice-versa.

This commit addresses both problems.