(maint) Ensure that sequences don't backtrack
Prior to this commit `reconcile-sequence-for-column!` had the potential
to set a sequence to a lower value than its current value in cases where
the most recent row had been deleted.

As of this commit, the function will never allow the sequence value to
be set lower than its current value.