(PE-18114) add lifetime protocol to pool
There have been issues with the connection pool being closed prior to
the initialization routines completing.  This results in some
transient failures in tests as well as actual application.

The lifetime functions allow the caller to block until the init
routines are complete.  An interface to query the state of the init
routine completion is added.  Finally an interface is provided to
attempt to cancel the init routine, if possible.

Many of the pool unit test functions were failing to close the pool
between tests.  with-open was applied to make sure the tests cleanup.