(TK-435) fns to create & destroy DB's & users
Add functions that check the existence of, create, and destroy both
databases and users. They all take a DB spec as their first argument.
The DB spec should connect to a database that's not the one being
created or destroyed, and the user it connects as must have the
appropriate permissions for the operation.