(maint) Silence spurious error logging for tests.
Previously, log4j logged at info level for everything besides the
jdbc-util.core namespace (which logged at debug level). This meant that
every test run had many spurious stack traces and errors logged which
seem to be problems but are actually expected. When there was an error,
it was difficult to pick out from all this noise.

In order to make it easy to find test failures and errors, configure
log4j to turn off logging for everything that's expected to log during
tests. Any failures or exceptions encountered will still be logged by
clojure.test.