(TK-435) Add input-sanitization fns.
Add `pg-sql-escape`, which takes an arbitrary string and returns it in
postgres's dollar-quoted format, which makes it safe to naively
interpolate into raw SQL strings. However, this can only be used in
places where postgres expects strings; it cannot be used to escape
identifiers such as the names of databases, users, or columns.

Add `safe-pg-identifier?`, for use when it's necessary to interpolate
identifiers into an SQL query. This predicate returns true when the
identifier is composed of a safe subset of characters: alphabetical
characters, digits, and the symbols '-' and '_', to the exclusion of all
other symbols.