(maint) Can create databases owned by other users
Previously, `create-db!` would throw an error when creating a database
owned by a different user than the user in the DB spec (unless the user
in the DB spec was a superuser). This is because postgres forbids users
from setting a database's owner to a role that they are not a member of.

Change `create-db!` to handle this case by having it check to see if the
DB spec user does have the required role membership. If not, it
temporarily grants the DB spec user the owner's role, creates the
database, and then revokes the role.