(PE-23367) update pglogical-ddl-wrapper to support multiples
As part of the migratus update, migratus now does migration of
its own schema_migrations table.  This involves two commands
used to upgrade the database.  For pglogical support, a
wrapper was created that made sure the commands sent were wrapped
in the proper pglogical commands.

The wrapper assumed a single command being sent to be wrapped.

This updates the wrapper to handle both single commands, and a
sequence of commands.  It separates the sequence of commands
with semicolons.