(main) Improve pool migration and initialization behavior
Previously, if an exception occurred while the pool was being created
it was logged, and connections were handed out normally.  This alters
the pool initialization behavior to:
* retry if transient failures occur during initialization
* log failures if non-transient failures are detected
* throw exceptions on getConnection attempts if the database failed to initialize correctly

This also increments the minor version number to 1.2 to indicate an interface
behavioral change.

Many test updates were made to improve the reliability of the tests.