(PE-24859) wait for safe migration exit prior to closing pool
In some circumstances, a close-after-ready was used.  The migrations
took longer than the timeout, which triggers the future cancellation.
Unfortunately, not waiting for the future cancel to have an impact prior
to closing the pool causes migratus to fail to clean up the migration
correctly, leaving it in a half-applied state.  This change causes
the pool to wait for the cancellation to have an impact, but also
applies a timeout to prevent a permanent wait scenario.