(PE-27018) uncompleted-migrations should have no side-effects
The "uncompleted-migrations" function has the side-effect of creating
the migration table if it doesn't exist.  In the case where we are
a replica system, this can cause issues with pglogical based replication.

This alters the uncompleted-migrations routine to first determine
if the database table exists, and if it does, determine what migrations
remain.  In the case that the table doesn't exist, it returns the
list of all the uncompleted migrations.

It adds tests for the table-exists? function, as well as a test
that (which fails prior to the fix) demonstrates that the migrations
table isn't present when waiting for migrations in a replica scenario.