(maint) Don't serialize nil to jsonb by default
Previously, calling (obj->jsonb nil) would result in the value
'null'::jsonb being written to the database. This confusingly bypasses a
NOT NULL constraint on the column (when we read and deserialize the
value, we unexpectedly still get back nil) and is almost never the
desired behavior.

This changes obj->jsonb to return nil when passed nil, with an option to
allow and serialize nil. This will cause attempts to write nil to a NOT
NULL jsonb column to properly result in an error.