(PE-38147) account for is_called when reconciling seqs
The next value of a postgres sequence will usually be last_value + 1,
but if is_called is false then the next value will just be last_value.

This would primarily have issues when reconciling a newly create seq,
such as from a replica, with a table with 1 row. Previously, because a
seq is initialized with the values (1, false) the guard WHEN would be
false if the max row ID was 1, but the next seq value would still be 1,
which would violate the primary key constraint.  Now when is_called is
false, and the max row value and last_value are equal, is_called will be
set to true in order to get the next value that will can be stored in
the table.